arm64: dts: qcom: Add device tree for Samsung Galaxy Tab A 9.7 (2015)
This dts adds support for wifi and lte variants of
Samsung Galaxy Tab A 9.7 (2015), a tablet released in 2015.

Add a device tree for gt510 with initial support for:

- SDHCI (internal and external storage)
- USB Device Mode
- UART (untested)
- Regulators
- WCNSS (WiFi/BT)
- GPIO keys

Most of the device tree is separated into
msm8916-samsung-gt5-common.dtsi as there is also gt58* devices
(Samsung Galaxy Tab A 8.0) that share a lot of hardware with gt510.

Signed-off-by: Jasper Korten <jja2000@gmail.com>
Co-developed-by: Nikita Travkin <nikita@trvn.ru>
Signed-off-by: Nikita Travkin <nikita@trvn.ru>